Fix SSH ready() method failing to connect
The ready() method originally spun up 5 SSH connections one after
another, but if the port on the host wasn't available, they would all
fail without waiting for the timeout.

Creating one connection at a time and waiting for it to fully time out
to start a new connection fixes this issue.